Add helper function
