+++ db/db.go
@@ -92,14 +92,3 @@ func convNilToBytes(byteArray []byte) []byte {
 	}
 	return byteArray
 }
-
-// Opt refers to an option item applied to DB initialization. The detailed option types vary depending on
-// the DB implementation, and it is the caller's responsibility to set the correct options.
-type Opt struct {
-	Name  string
-	Value interface{}
-}
-
-func (o Opt) String() string {
-	return fmt.Sprintf("%s=%v", o.Name, o.Value)
-}
